+++ src/payload/mod.rs
@@ -9,6 +9,7 @@ mod tensorial;
 
 pub use differentiable::Differentiable;
 pub use elementary::Elementary;
+pub use gemm::GemmTask;
 pub use shape::Shape;
 pub use tensor::Tensor;
 pub use tensorial::Tensorial;